feat: Created new section in side bar for frontend and backend projects
