--- frontend/src/DevelopersPortfolio/FullStackWorks/FullStackLandingPage.tsx
@@ -1,6 +1,0 @@
-
-export const FullStackLandingPage = () => {
-  return (
-    <div>FullStackLandingPage</div>
-  )
-}
