+++ frontend/src/Sidebar/Sidebar.tsx
@@ -3,8 +3,8 @@ import { TodaySection } from "./SidebarComponents/TodaySection";
 import { YesterdaySection } from "./SidebarComponents/YesterdaySection";
 import { PastThirtyDaysSection } from "./SidebarComponents/PastThirtyDaysSection";
 import { Footer } from "./SidebarComponents/Footer";
-import { FrontendSection } from "./FrontendSection";
-import { FullStackSectionSidebar } from "./SidebarComponents/FullStackProjectSection";
+import { Projects } from "./ProjectSection/Projects";
+
 export const Sidebar = () => {
 
 
@@ -19,8 +19,7 @@ export const Sidebar = () => {
               <div className="flex flex-col gap-2 pb-2 text-gray-100 text-sm">
                 <div>
                   <span>
-                    <FrontendSection/>
-                    <FullStackSectionSidebar/>
+                    <Projects/>
                     <TodaySection/>
                     <YesterdaySection/>
                     <PastThirtyDaysSection/>
